Separate calculateSalariesByGenderAndExperience function as a service
